[Q20-15] Added test cases for MapGraphToTripleController and AddNetworkApiController.
